requirements.txt
tensorflow
opencv-python
numpy
pillow
matplotlib
scikit-learn
streamlit
